The journey begins !
This is the bare minimum required to display native wayland clients
inside a winit window. It just displays windows' content, and no input
seems to be forwarded for now.

This is based around the `smallvil` example in `smithay`, and a big part of
the code is just copied from here.